Match hover state on snippets collection and snippets folders
