+++ enterprise/frontend/src/metabase-enterprise/snippets/components/SnippetCollectionMenu/SnippetCollectionMenu.tsx
@@ -6,7 +6,14 @@ import { useDispatch, useSelector } from "metabase/lib/redux";
 import type { SnippetCollectionMenuProps } from "metabase/plugins";
 import { addUndo } from "metabase/redux/undo";
 import { getUserIsAdmin } from "metabase/selectors/user";
-import { Box, Button, FixedSizeIcon, Icon, Menu } from "metabase/ui";
+import {
+  ActionIcon,
+  Box,
+  FixedSizeIcon,
+  Icon,
+  Menu,
+  Tooltip,
+} from "metabase/ui";
 
 export function SnippetCollectionMenu({
   collection,
@@ -51,20 +58,21 @@ export function SnippetCollectionMenu({
     }
   };
 
+  const optionsLabel = t`Snippet folder options`;
+
   return (
     <Box onClick={(e) => e.stopPropagation()}>
       <Menu position="bottom-end">
         <Menu.Target>
-          <Button
-            w={24}
-            h={24}
-            c="text-secondary"
-            size="compact-xs"
-            variant="subtle"
-            leftSection={<FixedSizeIcon name="ellipsis" size={16} />}
-            aria-label={t`Snippet collection options`}
+          <Tooltip
+            label={optionsLabel}
             onClick={(e) => e.stopPropagation()}
-          />
+            openDelay={1000}
+          >
+            <ActionIcon aria-label={optionsLabel} size="md">
+              <FixedSizeIcon name="ellipsis" size={16} c="text-medium" />
+            </ActionIcon>
+          </Tooltip>
         </Menu.Target>
         <Menu.Dropdown>
           {isArchived ? (
